Make the plugin work in the REPL
The solution is to try to load twice the impl we're interested in: First
by using a normal classloader, and then by using a classloader that is
aware of the path where the REPL outputs the classes.

This is the same solution as the one used in scalac, but less elegant
because we cannot directly use `findMacroClassloader`.

Maybe there exists a simple way to know if we're in the REPL?