Progress with whitebox parser macros
The fix for the parser is not commented anymore, because whitebox parser
macros seem to expand just fine in all the test cases :)

However, whitebox parser macro currently don't work well in REPL because
of a classpath problem. We need to get the "virtual classpath" of the
REPL to load macro implementation, but unfortunately this requires
access to data stored in Global, which we don't have in the macro
annotation.