Remove the need for `#` between arguments of parser macros
Until now parser the arguments of parser macros had to be put in
separate sets of parentheses, and to put a `#` between these.

This commit removes the need for this sign, but all the parentheses
and braces opened in the arguments of a parser macro application must
now be closed.